proyectos/air-bot/requirements.txt
# ===================================
# AIR-Bot - Dependencias de Python
# ===================================

# Bot de Telegram
python-telegram-bot>=20.7
python-telegram-bot[webhooks]>=20.7

# Google AI (New SDK)
google-genai

# Procesamiento de imágenes
Pillow>=10.2.0

# Variables de entorno
python-dotenv>=1.0.0

# Utilidades
requests>=2.31.0
aiohttp>=3.9.0

# Fechas y tiempo
python-dateutil>=2.8.2
